chatbot/pyproject.toml
[project]
name = "chatgzq-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faiss-cpu>=1.10.0",
    "fastapi[standard]>=0.115.12",
    "groq>=0.23.0",
    "pymupdf>=1.25.5",
    "python-dotenv>=1.1.0",
    "sentence-transformers>=4.1.0",
    "transformers>=4.51.3",
]
